This is to fix your issue with threshold value only accepting 1 digit. it now accepts minimum of 1 and maximum of 3 https://github.com/angelovstanton/Projects/pull/6 Our UI suite has > 255 tests. Because of the usage of bytes through the MSTest extender solution, failed tests in runs of > 255 would not retry. Replaced bytes with ints to allow for retry execution of these sets. https://github.com/angelovstanton/AutomateThePlanet/pull/2

+++ MSTest.Console/MSTest.Console.Extended/Infrastructure/MsTestTestRunProvider.cs
@@ -34,11 +34,11 @@ namespace MSTest.Console.Extended.Infrastructure
         public List<TestRunUnitTestResult> GetAllPassesTests(TestRun testRun)
         {
             List<TestRunUnitTestResult> results = new List<TestRunUnitTestResult>();
-            
+
             results = testRun.Results.ToList().Where(x => x.outcome.Equals("Passed")).ToList();
             return results;
         }
-        
+
         public void UpdatePassedTests(List<TestRunUnitTestResult> passedTests, List<TestRunUnitTestResult> allTests)
         {
             foreach (var currentTest in allTests)
@@ -56,11 +56,11 @@ namespace MSTest.Console.Extended.Infrastructure
             results = allTests.Where(x => !x.outcome.Equals("Passed")).ToList();
             return results;
         }
-      
+
         public void UpdateResultsSummary(TestRun testRun)
         {
-            testRun.ResultSummary.Counters.failed = (byte)testRun.Results.ToList().Count(x => x.outcome.Equals("Failed"));
-            testRun.ResultSummary.Counters.passed = (byte)testRun.Results.ToList().Count(x => x.outcome.Equals("Passed"));
+            testRun.ResultSummary.Counters.failed = (int)testRun.Results.ToList().Count(x => x.outcome.Equals("Failed"));
+            testRun.ResultSummary.Counters.passed = (int)testRun.Results.ToList().Count(x => x.outcome.Equals("Passed"));
             if ((int)testRun.ResultSummary.Counters.passed != testRun.Results.Length)
             {
                 testRun.ResultSummary.outcome = "Failed";
@@ -94,7 +94,7 @@ namespace MSTest.Console.Extended.Infrastructure
             {
                 result = ((double)failedTests.Count / (double)allTests.Count) * 100;
             }
-            
+
             return (int)result;
         }
     }
